!query affects patch
"querystring", which is a 3rd party app used to deconstruct URL's or in this case, the argument. In URL's you need to use %2B instead of "+" for proper handling.  When we were sending the + to query string it was returning a blank space, so technically the + was always being ignored.  by adding the .replace to the queryParam (before being processed by querystring) the output from querystring will include +'s which of course makes all my previous patches now work properly.