!query affects= patch
Was running into the 2000 character limit on "!query affects=spell slots by +1 at level 6"    In order to prevent this in the future, I have wrote some code that detects the size of the message variable as its building it and if it gets over 1900, it starts a 2nd message and then sends them both.  This won't happen very often and I think the only reason it happened this time is because creators of +1 6th slot items tend to put more object names on the item, so "breast plate breastplate dragon dragonscale scale" etc, where as lower quality items would be like "breastplate dragonscale"  I think that's why the threshold broke 2000 and wasn't accounted for.  Either way, this code fixes that so that the user would still get the data they requested rather than just "there were 49 items."