CreateUpdateLore function - patch to fix #29
The issue here was a variable that has a numerical value of 0, when evaluated in parenthesis will evaluate as false.  Since we need a value of 0 to evaluate as true, we need to convert it to a string before attempting the evaluation.  i added the conversion for all the variables to prevent other issues, such as 0 apply, or 0 value, or basically anything that can be the number 0 as a value.  I had to check for null before attempting to convert using .toString() due to errors thrown when trying to run .toString() against a null value.

I learned how to fix this from here:
https://stackoverflow.com/questions/7615214/in-javascript-why-is-0-equal-to-false-but-when-tested-by-if-it-is-not-fals
https://medium.com/@sgobinda007/dealing-with-javascript-falsy-values-d75a2f1b1c90